moved files, added test
