+++ src/Dialect/ONNX/CMakeLists.txt
@@ -22,6 +22,7 @@ add_onnx_mlir_library(OMONNXOps
   ONNXDialect.cpp
   ONNXDimAnalysis.cpp
   ONNXOps.cpp
+  ONNXOps/FusionOpChain.cpp
   ONNXOps/OpHelper.cpp
   ONNXOps/Canonicalize.cpp
   ONNXOps/ShapeHelper.cpp
