+++ src/Transform/EliminateWriteOnlyAlloc.cpp
